+++ index.html
@@ -234,6 +234,7 @@
       </section>
     </template>
 <!-- <script src="/js/functions.js"></script> -->
-<script src="/js/main.js" defer></script>
+<script src="/js/main.js" type="module" defer></script>
+<!-- <script src="/js/data.js" type="module"></script> -->
 </body>
 </html>
